ttlock-sdk-py/requirements.txt
aioesphomeapi~=44.8.0
pycryptodome~=3.23.0
